services/ml-surrogate/requirements.txt
# ML Surrogate Service Dependencies

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime-gpu>=1.15.0  # GPU acceleration

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
redis>=4.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Logging
python-json-logger>=2.0.0
